+++ SunSmart/Main/Space/Controller/SpaceViewController.swift
@@ -8,17 +8,35 @@
 import UIKit
 import NordicSigMeshSDK
 import CoreBluetooth
+import SwiftyJSON
 
 /// 空间内菜单选择修改通知
 let spaceMenuIndexChangeNotificaitonName = "spaceMenuIndexChangeNotificaiton"
+/// 空间内数据修改通知
+/// 添加设备、编辑设备、删除设备、修复设备
+/// 添加组、编辑组（基本数据 、 添加/删除设备、profile、校准、动能开关）、删除组
+/// 添加场景、编辑场景（基本数据、添加/删除组、修改组参数）、删除场景
+/// 添加定时、编辑定时（名称、target、enable…）、删除定时
+let spaceDataChangedNotificaitonName = "spaceDataChangedNotificaiton"
+
+/// space修改数据类型
+enum SpaceChangeDataType {
+    /// 设备数据（包含device、group/scene等配置数据-与设备数据交互）
+    case device
+    /// 通用数据（group/scene等配置数据-无设备数据交互）
+    case common
+}
 
 class SpaceViewController: WMPageController {
 
+    var site: SiteData!
     let space: SpaceData
     /// 删除空间回调
     var deleteSpaceCallback: (()->Void)?
     /// 是否已加载完成网络数据
     private var loadNetworkData: Bool = false
+    /// 退出页面同步space中
+    private var exitSyncSpace: Bool = false
     
     lazy var mainMenuView: SpaceMenuView = {
         let menuView = SpaceMenuView(frame: CGRect(x: 0, y: kNavigationHeight, width: self.view.width, height: SCRYFrom(46)))
@@ -55,7 +73,7 @@ class SpaceViewController: WMPageController {
         title = space.name
         view.backgroundColor = Background_Color
         menuView?.backgroundColor = .white
-        view.addSubview(mainMenuView)
+        self.view.addSubview(self.mainMenuView)
         
         navigationItem.rightBarButtonItem = UIBarButtonItem(image: UIImage(named: "more_vertical")?.withRenderingMode(.alwaysOriginal), style: .done, target: self, action: #selector(moreClick))
 
@@ -66,35 +84,43 @@ class SpaceViewController: WMPageController {
         MeshLibManager.manager.publishModeloOnly = true
         MeshLibManager.manager.groupSubscriptionModelIDs = [.genericOnOffServerModelId, .lightLightnessServerModelId, .genericLevelServerModelId, .lightCTLTemperatureServerModelId, .lightCTLServerModelId, .sensorServerModelId, .lightLCServerModelId]
         checkBluetoothState()
-        // 读取网络数据
-        XWHUDManager.showCustomHUD(withMessage: nil, isWindow: false, afterDelay: 2)
-        DispatchQueue.global().async {
-            MeshLibManager.manager.setMeshNetworkConnected(meshUUID: self.space.meshUUID, subNetwork: self.space.meshNetworkKey)
-            
-            if let manager = MeshLibManager.manager.meshNetworkManager {
-                self.space.meshManager = manager
-                manager.loadExtensionData {[weak self] in
-                    guard let self = self else { return }
-//                    XWHUDManager.hideInView(with: self.view)
-                    self.loadNetworkData = true
-                    self.reloadData()
-                    self.configurationFlowGuidance()
-                }
-            }
-        }
    
         // 添加通知监听
         addNotificaiton()
+        // 获取space数据
+        setNetworkConnected()
+//        loadSpaceReqeust()
+    }
+    
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
         
+        CloudSynchronizationManager.shared.delegate = self
+        (self.navigationController as? NavigationViewController)?.navigationDelegate = self
+        
+    }
+    
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        
+        updateSyncState()
+    }
+    
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        ConfigurationFlowGuidanceView.current()?.hide()
     }
     
     /// 配置引导
     private func configurationFlowGuidance() {
         
+        guard !exitSyncSpace else {
+            return
+        }
         // 判断是否空的空间，进行引导配置流程
-        if space.isEmpty {
-            DispatchQueue.main.asyncAfter(wallDeadline: .now() + 0.5) {[weak self] in
-                guard let self = self else { return }
+        if view.window != nil && space.permission != .visitor && space.isEmpty {
+//            DispatchQueue.main.asyncAfter(wallDeadline: .now() + 0.5) {[weak self] in
+//                guard let self = self else { return }
                 ConfigurationFlowGuidanceView(continueBack: {[weak self] in
                     guard let self = self else { return }
                     // 进入引导配置流程
@@ -104,7 +130,7 @@ class SpaceViewController: WMPageController {
                     self.present(navVc, animated: true)
                     self.selectIndex = 1
                 }).show()
-            }
+//            }
         }
         
     }
@@ -146,17 +172,90 @@ class SpaceViewController: WMPageController {
         NotificationCenter.default.addObserver(forName: .init(schedulesRefreshNotificationName), object: nil, queue: nil) {[weak self] _ in
             self?.updateSpaceData()
         }
-        
+        // 空间内菜单选择修改通知
         NotificationCenter.default.addObserver(forName: .init(spaceMenuIndexChangeNotificaitonName), object: nil, queue: nil) {[weak self] notification in
             guard let self = self, let selectIndex = notification.object as? Int, selectIndex >= 0 && selectIndex < SpaceMenuView.defalutItems.count else { return }
             self.selectIndex = Int32(selectIndex)
         }
+        // 空间内数据更新通知
+        NotificationCenter.default.addObserver(forName: .init(spaceDataChangedNotificaitonName), object: nil, queue: nil) {[weak self] notification in
+            guard let self = self, let type = notification.object as? SpaceChangeDataType else { return }
+            self.space.lastUpdate = Int64(Date().timeIntervalSince1970)
+            self.space.save()
+            switch type {
+            case .device:
+                CloudSynchronizationManager.shared.addSynchronizationHandle(operation: .syncSpace(space: self.space), level: .promptly)
+            case .common:
+                CloudSynchronizationManager.shared.addSynchronizationHandle(operation: .syncSpace(space: self.space), level: .slow)
+            }
+        }
         
-        // 日程列表更新通知
-//        NotificationCenter.default.addObserver(forName: .init(schedulesRefreshNotificationName), object: nil, queue: nil) {[weak self] _ in
-//            self?.updateSpaceData()
-//        }
     }
+    
+    /// 获取网络数据+网络连接
+    private func setNetworkConnected() {
+        // 读取网络数据
+        XWHUDManager.showCustomHUD(withMessage: nil, isWindow: false, afterDelay: 2)
+        DispatchQueue.global().async {
+            MeshLibManager.manager.setMeshNetworkConnected(meshUUID: self.space.meshUUID, subNetwork: self.space.meshNetworkKey)
+            if let manager = MeshLibManager.manager.meshNetworkManager {
+                self.space.meshManager = manager
+                manager.loadExtensionData {[weak self] in
+                    guard let self = self else { return }
+//                    XWHUDManager.hideInView(with: self.view)
+                    self.loadNetworkData = true
+                    self.reloadData()
+                    self.configurationFlowGuidance()
+                }
+            }
+        }
+    }
+    
+    // MARK: - Request
+    /// 获取space数据
+    private func loadSpaceReqeust() {
+        
+        XWHUDManager.showCustomHUD(withMessage: nil, isWindow: false)
+        NetworkRequest.shared.request(.spaceInfo(siteId: space.siteId, spaceId: space.id, password: "")) {[weak self] result in
+            guard let self = self else { return }
+            XWHUDManager.hideInView(with: self.view)
+            switch result {
+            case .success(let response):
+                if let spaceData = JSON(response)["data"].dictionaryObject {
+                    Task {
+                        await self.space.update(spaceJsonData: spaceData)
+                        self.title = self.space.name
+                        self.space.save()
+                    }
+                }
+            case .failure(_):
+                break
+            }
+            self.setNetworkConnected()
+        }
+        
+    }
+    
+    /// 删除space网络请求
+    private func deleteSpaceRequest() {
+        
+        XWHUDManager.showCustomHUD(withMessage: "deleting".localizedString, isWindow: true)
+        NetworkRequest.shared.request(.siteDelete(siteId: self.site.id)) {[weak self] result in
+            XWHUDManager.hide()
+            switch result {
+            case .success(_):
+                // 删除本地数据
+                self?.space.delete()
+                self?.navigationController?.popViewController(animated: true)
+                self?.deleteSpaceCallback?()
+                NotificationCenter.default.post(name: .init(rawValue: SitesDataRefreshNotifiacationName), object: nil)
+                
+            case .failure(let error): // 删除失败，无网络/space存在编辑者
+                XWHUDManager.showTipHUD(error.localizedDescription, isLineFeed: true)
+            }
+        }
+    }
+    
     
     /// 更新空间缓存数据
     private func updateSpaceData() {
@@ -242,14 +341,30 @@ class SpaceViewController: WMPageController {
             return
         }
         
-        MenuPopView.show(items: [
-            .init(icon: UIImage(named: "edit"), title: "edit_space".localizedString, tapItemBack: {[weak self] item in
+        var items: [MenuPopView.MenuItem] = []
+        
+        if space.spaceOperates.contains(.edit) {
+            items.append(.init(icon: UIImage(named: "menu_edit"), title: "edit".localizedString, tapItemBack: {[weak self] _ in
                 self?.editSpace()
-            }),
-            .init(icon: UIImage(named: "menu_delete"), title: "delete_space".localizedString, tapItemBack: {[weak self] item in
+            }))
+        }
+        if space.spaceOperates.contains(.delete) {
+            items.append(.init(icon: UIImage(named: "menu_delete"), title: "delete".localizedString, tapItemBack: {[weak self] _ in
                 self?.deleteSpace()
-            }),
-        ], anchorPoint: CGPoint(x: view.width - 18 - 15, y: kNavigationHeight), menuWidth: SCRXFrom(154))
+            }))
+        }
+        if space.spaceOperates.contains(.shareEditor) || space.spaceOperates.contains(.shareVisitor) {
+            items.append(.init(icon: UIImage(named: "menu_share"), title: "share".localizedString, tapItemBack: {[weak self] _ in
+                self?.shareSpace()
+            }))
+        }
+        if space.spaceOperates.contains(.exit) {
+            items.append(.init(icon: UIImage(named: "menu_unbind"), title: "unbind".localizedString, tapItemBack: {[weak self] _ in
+                self?.unbindSpace()
+            }))
+        }
+        
+        MenuPopView.show(items: items, anchorPoint: CGPoint(x: view.width - 18 - 15, y: kNavigationHeight), menuWidth: SCRXFrom(154))
     }
     
     /// 编辑空间
@@ -265,11 +380,13 @@ class SpaceViewController: WMPageController {
             return SpaceData.isTautonym(spaceName: name, siteId: self.space.siteId) && name != self.space.name
         }
         vc.doneCallback = {[weak self] (name, imageId) in
-            guard let self = self else { return }
+            guard let self = self else { return true }
             self.space.name = name
             self.space.imageId = imageId + 1
             self.space.save()
             self.title = name
+            CloudSynchronizationManager.shared.addSynchronizationHandle(operation: .syncSpace(space: self.space), level: .normal)
+            return true
         }
         present(NavigationViewController(rootViewController: vc), animated: true)
     }
@@ -287,14 +404,69 @@ class SpaceViewController: WMPageController {
                 if self.space.deviceCount > 0 {
                     XWHUDManager.showErrorTipHUD("site_delete_fail".localizedString)
                 }else { // 空间未存在设备，删除成功
-                    self.space.delete()
-                    self.navigationController?.popViewController(animated: true)
-                    self.deleteSpaceCallback?()
+                    // 提交到云端需要网络才能删除
+                    if self.space.uploadCloud {
+                        self.deleteSpaceRequest()
+                    }else { // 只存在于本地，删除数据
+                        self.space.delete()
+                        self.navigationController?.popViewController(animated: true)
+                        self.deleteSpaceCallback?()
+                        NotificationCenter.default.post(name: .init(rawValue: SitesDataRefreshNotifiacationName), object: nil)
+                    }
+          
                 }
-                NotificationCenter.default.post(name: .init(rawValue: SitesDataRefreshNotifiacationName), object: nil)
             }
         })]).show()
         
+    }
+    
+    /// 分享space
+    private func shareSpace() {
+        let vc = SharingSettingViewController(type: .space(site: self.site, space: self.space))
+        present(NavigationViewController(rootViewController: vc), animated: true)
+    }
+    
+    /// 解绑space
+    private func unbindSpace() {
+        
+    }
+    
+    
+    /// 更新同步状态
+    private func updateSyncState() {
+        
+        if view.window != nil, let state = CloudSynchronizationManager.shared.getSpaceCurrentSyncState(space)?.state {
+            switch state {
+            case .inProgress:
+                self.showNavigationBarLoading()
+            case .successful:
+                self.showNavigationBarSuccessful()
+            case .failure:
+                self.showNavigationBarFailure {[weak self] in
+                    // 点击失败图标
+                    
+                }
+            default:
+                break
+            }
+        }
+    }
+    
+    /// 退出页面立即同步space数据
+    private func promptlySyncSpace() {
+        
+        ConfigurationFlowGuidanceView.current()?.hide()
+        
+        CloudSynchronizationManager.shared.addSynchronizationHandle(operation: .syncSpace(space: space), level: .promptly)
+        exitSyncSpace = true
+        XWHUDManager.showCustomHUD(withMessage: "syncing_data".localizedString, isWindow: true)
+    }
+    /// 展示同步space数据失败页面
+    private func showSyncSpaceFailedAlert() {
+        
+        SRAlertView(title: "notification".localizedString, message: "space_sync_failed_message".localizedString, actions: [SRAlertAction(title: "confirm".localizedString, actionHandler: {[weak self] _ in
+            self?.navigationController?.popViewController(animated: true)
+        })]).show()
     }
     
 }
@@ -350,4 +522,65 @@ extension SpaceViewController {
 //        return index < 3
     }
     
+}
+
+extension SpaceViewController: NavigationViewControllerDelegate {
+    
+    /// 点击返回item回调
+    func navigationController(_ navigationController: NavigationViewController, backItemAction showViewController: UIViewController) {
+        
+        guard space.needUploadCloud else {
+            self.navigationController?.popViewController(animated: true)
+            return
+        }
+        promptlySyncSpace()
+    }
+    
+    /// pop手势begin回调，返回是否可以pop
+    func navigationController(_ navigationController: NavigationViewController, gestureRecognizerShould gestureRecognizer: UIGestureRecognizer) -> Bool {
+        // 无网络并且更新了数据
+        guard selectIndex == 0 else {
+            return false
+        }
+        guard space.needUploadCloud else {
+            return true
+        }
+        promptlySyncSpace()
+        return false
+    }
+}
+
+extension SpaceViewController: CloudSynchronizationManagerDelegate {
+    
+    /// 开始同步数据回调
+    /// - Parameters:
+    ///   - manager: 同步管理
+    ///   - handle: 同步数据操作
+    func cloudSynchManager(_ manager: CloudSynchronizationManager, didStartSync handle: CloudSynchronizationHandle) {
+        updateSyncState()
+    }
+    
+    /// 同步数据成功回调
+    /// - Parameters:
+    ///   - manager: 同步管理
+    ///   - handle: 同步数据操作
+    func cloudSynchManager(_ manager: CloudSynchronizationManager, didSyncFinished handle: CloudSynchronizationHandle) {
+        updateSyncState()
+        if exitSyncSpace {
+            XWHUDManager.hide()
+            navigationController?.popViewController(animated: true)
+        }
+    }
+    
+    /// 同步数据失败回调
+    /// - Parameters:
+    ///   - manager: 同步管理
+    ///   - handle: 同步数据操作
+    func cloudSynchManager(_ manager: CloudSynchronizationManager, didSyncFailure handle: CloudSynchronizationHandle, error: NetworkApiError) {
+        updateSyncState()
+        if exitSyncSpace { // 提示数据同步失败
+            XWHUDManager.hide()
+            showSyncSpaceFailedAlert()
+        }
+    }
 }
